# swaggerui/swagger.yaml
swagger: '2.0'
info:
  description: TomoDex API Document
  version: 1.0.0
  title: TomoDex APIs
schemes:
  - https
basePath: /api
...
paths:
  '/account/{userAddress}':
    get:
      tags:
        - accounts
      summary: Find account by user address
      description: Returns a single account
      operationId: handleGetAccount
      produces:
        - application/json
      parameters:
        - name: userAddress
          in: path
          description: Address of user to return
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Account'
        '400':
          description: Invalid Address
        '404':
          description: Account not found
  '/account/{userAddress}/{tokenAddress}':
    get:
      tags:
        - accounts
      summary: '[Deprecated] Find account''s token balance by user address and token address'
      description: Returns an object contains token balance of user
      operationId: handleGetAccountTokenBalance
      deprecated: true
      produces:
        - application/json
      parameters:
        - name: userAddress
          in: path
          description: Address of user to find token balance
          required: true
          type: string
        - name: tokenAddress
          in: path
          description: Address of token
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/TokenBalance'
        '400':
          description: Invalid Address
        '404':
          description: Account not found
  /tokens:
    get:
      tags:
        - tokens
      summary: Finds all tokens
      description: Return all tokens in an array
      operationId: HandleGetTokens
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Token'
        '500':
          description: Internal Server Error
    post:
      tags:
        - tokens
      summary: Create new token
      description: Returns newly created token
      operationId: HandleCreateToken
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Token object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/Token'
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            $ref: '#/definitions/Token'
        '400':
          description: Invalid payload
        '500':
          description: Internal Server Error
  /tokens/base:
    get:
      tags:
        - tokens
      summary: Finds all base tokens
      description: Return all base tokens in an array
      operationId: HandleGetBaseTokens
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Token'
        '500':
          description: Internal Server Error
  /tokens/quote:
    get:
      tags:
        - tokens
      summary: Finds all quote tokens
      description: Return all quote tokens in an array
      operationId: HandleGetQuoteTokens
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Token'
        '500':
          description: Internal Server Error
  '/tokens/{address}':
    get:
      tags:
        - tokens
      summary: Retrieve the token information corresponding to an address
      description: Return token object
      operationId: HandleGetToken
      produces:
        - application/json
      parameters:
        - name: address
          in: path
          description: Token address
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            $ref: '#/definitions/Token'
        '400':
          description: Invalid Address
        '500':
          description: Internal Server Error
  /pairs:
    get:
      tags:
        - pairs
      summary: Finds all pairs
      description: Return all pairs in an array
      operationId: HandleGetPairs
      produces:
        - application/json
      parameters:
        - name: baseToken
          in: query
          description: Base token address
          required: true
          type: string
        - name: quoteToken
          in: query
          description: Quote token address
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Pair'
        '500':
          description: Internal Server Error

    post:
      tags:
        - pairs
      summary: Create new pair
      description: Returns newly created pair
      operationId: HandleCreatePair
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Pair object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/Pair'
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            $ref: '#/definitions/Pair'
        '400':
          description: '*** Parameter missing'
        '500':
          description: Internal Server Error
  /pair:
    get:
      tags:
        - pairs
      summary: Retrieve the pair information corresponding to a baseToken and a quoteToken
      description: Multiple status values can be provided with comma separated strings
      operationId: HandleGetPair
      produces:
        - application/json
      parameters:
        - name: baseToken
          in: query
          description: Base token address
          required: true
          type: string
        - name: quoteToken
          in: query
          description: Quote token address
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            $ref: '#/definitions/Pair'
        '400':
          description: baseToken Parameter missing
        '500':
          description: Internal Server Error
  /pair/data:
    get:
      tags:
        - pairs
      summary: Retrieve pair data corresponding to a baseToken and quoteToken
      description: Multiple status values can be provided with comma separated strings
      operationId: HandleGetPairData
      produces:
        - application/json
      parameters:
        - name: baseToken
          in: query
          description: Base token address
          required: true
          type: string
        - name: quoteToken
          in: query
          description: Quote token address
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            $ref: '#/definitions/PairData'
        '400':
          description: baseToken Parameter missing
        '500':
          description: Internal Server Error
  /pairs/data:
    get:
      tags:
        - pairs
      summary: Retrieve all pair data
      description: Multiple status values can be provided with comma separated strings
      operationId: HandleGetPairsData
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            $ref: '#/definitions/PairData'
        '400':
          description: baseToken Parameter missing
        '500':
          description: Internal Server Error
...
definitions:
  Account:
    example:
      address: '0xF7349C253FF7747Df661296E0859c44e974fb52E'
    type: object
    properties:
      id:
        type: string
        readOnly: true
      address:
        type: string
      tokenBalances:
        type: object
        properties:
          address:
            type: string
          symbol:
            type: string
          balance:
            type: string
          availableBalance:
            type: string
          inOrderBalance:
            type: string
      isBlocked:
        type: boolean
      createdAt:
        type: string
        format: date-time
        readOnly: true
      updatedAt:
        type: string
        format: date-time
        readOnly: true
  TokenBalance:
    type: object
    properties:
      address:
        type: string
      symbol:
        type: string
      balance:
        type: string
      availableBalance:
        type: string
      inOrderBalance:
        type: string
  Token:
    type: object
    properties:
      id:
        type: string
        readOnly: true
      name:
        type: string
      symbol:
        type: string
      address:
        type: string
      image:
        $ref: '#/definitions/Image'
        readOnly: true
      contractAddress:
        type: string
      decimals:
        type: integer
        format: int32
        readOnly: true
      active:
        type: boolean
      listed:
        type: boolean
        readOnly: true
      quote:
        type: boolean
      makeFee:
        type: string
      takeFee:
        type: string
      usd:
        type: string
        readOnly: true
      createdAt:
        type: string
        format: date-time
        readOnly: true
      updatedAt:
        type: string
        format: date-time
        readOnly: true
  Pair:
    type: object
    properties:
      id:
        type: string
        readOnly: true
      baseTokenSymbol:
        type: string
      baseTokenAddress:
        type: string
      baseTokenDecimals:
        type: integer
        format: int32
        readOnly: true
      quoteTokenSymbol:
        type: string
      quoteTokenAddress:
        type: string
      quoteTokenDecimals:
        type: integer
        format: int32
        readOnly: true
      listed:
        type: boolean
        readOnly: true
      active:
        type: boolean
      rank:
        type: integer
        format: int32
        readOnly: true
      makeFee:
        type: string
        readOnly: true
      takeFee:
        type: string
        readOnly: true
      createdAt:
        type: string
        format: date-time
        readOnly: true
      updatedAt:
        type: string
        format: date-time
        readOnly: true
  PairID:
    type: object
    properties:
      pairName:
        type: string
      baseToken:
        type: string
      quoteToken:
        type: string
  PairData:
    type: object
    properties:
      pair:
        $ref: '#/definitions/PairID'
      open:
        type: string
      high:
        type: string
      low:
        type: integer
        format: int32
      close:
        type: string
      volume:
        type: string
      count:
        type: string
      timestamp:
        type: string
      orderVolume:
        type: string
      orderCount:
        type: string
      averageOrderAmount:
        type: string
      averageTradeAmount:
        type: string
      askPrice:
        type: string
      bidPrice:
        type: string
      price:
        type: string
      rank:
        type: integer
        format: int32
  Image:
    type: object
    properties:
      url:
        type: string
      meta:
        type: object
        additionalProperties: true
...